pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-geo-spy"
version = "2.0.0"
description = "Production-grade multi-agent geolocation system"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    # Core
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "sse-starlette>=2.0.0",

    # Agentic orchestration
    "langgraph>=0.2.0",

    # Logging & retry
    "loguru>=0.7.2",
    "tenacity>=8.2.3",

    # LLM / AI
    "openai>=1.30.0",

    # ML Models
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.40.0",
    "open-clip-torch>=2.24.0",

    # Image processing
    "pillow>=10.0.0",
    "piexif>=1.1.3",
    "numpy>=1.24.0",

    # Browser / web
    "browser-use>=0.11.0",
    "curl-cffi>=0.7.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.0",
    "httpx>=0.27.0",

    # Geo
    "shapely>=2.0.0",
    "geopy>=2.4.0",
    "overpy>=0.7",
    "geoclip>=1.2.0",
    "playwright>=1.40.0",
    "reverse-geocoder>=1.5.1",

    # Data
    "orjson>=3.9.0",

    # Calibration
    "scikit-learn>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

stealth = [
    "browser-use-undetected>=0.1.0",
    "camoufox>=0.4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
